New GuildEventLogger class
This allows anything to report events to a specific guild by using the
instance of the class after setting it up using it's Initialize method.

The GuildEventLoggerCommands serves as an interface between the Discord
Server admins and the logger itself.

It still misses some functionality like sending a message to all the
tracked guild but I guess it's fine for a first flesh out