Port the Addons specific commands to SlashCommands
This patch uses the new SlashCommands and implements the addon listing, the addition and deletion of addons, it also makes use of some SlashCommands specific features like auto completion, user restrictions and others. I suppose the idea would be to eventually have all of the commands ported to SlashCommands but this is a start.