AddonsCtl AddMb & AddDl separation / improvements
This patch separates the addon add command in 2:
- One of them takes a SRB2 message board link
- One of them takes a direct link to the Addon

The addon download from the message board has received a significant improvement from the fact that it can now ask what addon file to download if there are multiple and not just download the HTML file.

Some parts have also been extracted to methods to have more compact and readabke code.

Finally, this adds the HtmlAgilityPack nuget package to parse the potential html page for a message board addon and removes the unused Newtonsoft.Json package