Cleaned up and improved code for the SlashCommands
This patch reduces the SlashCommands to only 4 by using a single command for each action and not separating the levels, characters and 2.1 addons.
The only real downside to this is when using the delete command, the way Discord caches the AutoComplete choices is quite unfortunate in this case but it's good enough in my opinion.