Addons detection improvements
This changes the behavior of the addons detection when somebody uses the reptogif / mp4 command.
In addition to checking for the file name contained in the replay file, it'll now also check for the MD5 hash against all of our addons on the server.
There is a FileSystemWatcher that will watch for any changes in the addons directory and reload all of the file hashes when a change occurs.
This means that we won't be computing the hashes each time a replay is sent but only when there are changes made to the addons.